Separación de menús, arreglo de logout, NoticiaController
