+++ src/main/java/com/uah/ismael/portal_formula1/service/impl/NoticiaServiceImpl.java
@@ -10,11 +10,13 @@ import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
+@Service
 public class NoticiaServiceImpl implements NoticiaService {
 
     private final ModelMapper modelMapper;
