--- src/main/resources/templates/admin/admin_home.html
@@ -1,47 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Admin Home</title>
-    <link rel="stylesheet" type="text/css" th:href="@{/css/bootstrap.min.css}">
-</head>
-<body>
-<div class="container">
-    <div class="header text-center my-4">
-        <div class="logo-container">
-            <a href="/" th:href="@{/}">
-                <img src="/images/logo.png" alt="Logo" class="logo img-fluid">
-            </a>
-            <h1>Administrador <span class="margen" th:text="${nombreUsuario}"></span></h1>
-        </div>
-    </div>
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <ul class="navbar-nav mr-auto">
-            <li class="nav-item"><a class="nav-link" th:href="@{/admin/show_users}">Usuarios</a></li>
-            <li class="nav-item dropdown">
-                <a href="javascript:void(0)" class="nav-link dropdown-toggle" data-toggle="dropdown">Noticias</a>
-                <div class="dropdown-menu">
-                    <a class="dropdown-item">Ver Noticias</a>
-                    <a class="dropdown-item">Crear Noticia</a>
-                </div>
-            </li>
-            <li class="nav-item dropdown">
-                <a href="javascript:void(0)" class="nav-link dropdown-toggle" data-toggle="dropdown">Votaciones</a>
-                <div class="dropdown-menu">
-                    <a class="dropdown-item" th:href="@{/admin/activate_users}">Ver Votaciones</a>
-                    <a class="dropdown-item" th:href="@{/admin/delete_users}">Crear Votación</a>
-                </div>
-            </li>
-            <li class="nav-item"><a class="nav-link" th:href="@{/logout}">Cerrar Sesión</a></li>
-        </ul>
-    </nav>
-    <div class="container my-4">
-        <!-- Contenido adicional aquí -->
-    </div>
-    <div class="footer text-center mt-4">
-        <p>&copy; Curso 24/25 Portal Fórmula 1 - Ismael Blázquez</p>
-    </div>
-</div>
-<script th:src="@{/js/bootstrap.min.js}"></script>
-</body>
-</html>
