+++ src/main/resources/templates/fragmentos/footer.html
@@ -1,0 +1,4 @@
+<!-- src/main/resources/templates/fragmentos/footer.html -->
+<div class="footer text-center mt-4" th:fragment="footer">
+    <p>&copy; Curso 24/25 Portal Fórmula 1 - Ismael Blázquez</p>
+</div>
