--- src/main/resources/templates/responsable/responsable_home.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Responsable de Equipo Home</title>
-    <link rel="stylesheet" type="text/css" th:href="@{/css/bootstrap.min.css}">
-</head>
-<body>
-<div class="container">
-    <div class="header text-center my-4">
-        <div class="logo-container">
-            <a href="/" th:href="@{/}">
-                <img src="/images/logo.png" alt="Logo" class="logo img-fluid">
-            </a>
-            <h1>Responsable de Equipo <span class="margen" th:text="${nombreUsuario}"></span></h1>
-        </div>
-    </div>
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <ul class="navbar-nav mr-auto">
-            <li class="nav-item"><a class="nav-link" th:href="@{/noticias}">Noticias</a></li>
-            <li class="nav-item"><a class="nav-link" th:href="@{/votaciones}">Votaciones</a></li>
-            <li class="nav-item"><a class="nav-link" th:href="@{/logout}">Iniciar Sesión</a></li>
-        </ul>
-    </nav>
-    <div class="container my-4">
-        <!-- Contenido adicional aquí -->
-    </div>
-    <div class="footer text-center mt-4">
-        <p>&copy; Curso 24/25 Portal Fórmula 1 - Ismael Blázquez</p>
-    </div>
-</div>
-<script th:src="@{/js/bootstrap.min.js}"></script>
-</body>
-</html>
