+++ src/main/resources/templates/usuario/show_users.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity5" lang="es">
 <head>
     <meta charset="UTF-8">
     <title>Mostrar Usuarios</title>
@@ -56,52 +57,26 @@
 </head>
 <body>
 <div class="container">
-    <div class="header text-center my-4">
-        <div class="logo-container">
-            <a href="/" th:href="@{/}">
-                <img src="/images/logo.png" alt="Logo" class="logo img-fluid">
-            </a>
-            <h1>Usuarios</h1>
-        </div>
-    </div>
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <ul class="navbar-nav mr-auto">
-            <li class="nav-item"><a class="nav-link" th:href="@{/admin/show_users}">Usuarios</a></li>
-            <li class="nav-item dropdown">
-                <a href="javascript:void(0)" class="nav-link dropdown-toggle" data-toggle="dropdown">Noticias</a>
-                <div class="dropdown-menu">
-                    <a class="dropdown-item">Ver Noticias</a>
-                    <a class="dropdown-item">Crear Noticia</a>
-                </div>
-            </li>
-            <li class="nav-item dropdown">
-                <a href="javascript:void(0)" class="nav-link dropdown-toggle" data-toggle="dropdown">Votaciones</a>
-                <div class="dropdown-menu">
-                    <a class="dropdown-item" th:href="@{/admin/activate_users}">Ver Votaciones</a>
-                    <a class="dropdown-item" th:href="@{/admin/delete_users}">Crear Votación</a>
-                </div>
-            </li>
-            <li class="nav-item"><a class="nav-link" th:href="@{/logout}">Cerrar Sesión</a></li>
-        </ul>
-    </nav>
+    <div th:replace="~{fragmentos/loguedMenu :: menu('Usuarios', '')}"></div>
+
     <div class="container my-4">
         <table class="table table-striped">
             <thead>
             <tr>
                 <th>
-                    <a th:href="@{/admin/show_users(page=${currentPage}, size=${userPage.size}, sortField='nombreUsuario', sortDir=${sortField == 'nombreUsuario' ? reverseSortDir : 'asc'})}">
+                    <a th:href="@{/show_users(page=${currentPage}, size=${userPage.size}, sortField='nombreUsuario', sortDir=${sortField == 'nombreUsuario' ? reverseSortDir : 'asc'})}">
                         Nombre de Usuario
                         <span th:if="${sortField == 'nombreUsuario'}" th:text="${sortDir == 'asc' ? '▲' : '▼'}"></span>
                     </a>
                 </th>
                 <th>
-                    <a th:href="@{/admin/show_users(page=${currentPage}, size=${userPage.size}, sortField='rols', sortDir=${sortField == 'rols' ? reverseSortDir : 'asc'})}">
+                    <a th:href="@{/show_users(page=${currentPage}, size=${userPage.size}, sortField='rols', sortDir=${sortField == 'rols' ? reverseSortDir : 'asc'})}">
                         Roles
                         <span th:if="${sortField == 'rols'}" th:text="${sortDir == 'asc' ? '▲' : '▼'}"></span>
                     </a>
                 </th>
                 <th>
-                    <a th:href="@{/admin/show_users(page=${currentPage}, size=${userPage.size}, sortField='activo', sortDir=${sortField == 'activo' ? reverseSortDir : 'asc'})}">
+                    <a th:href="@{/show_users(page=${currentPage}, size=${userPage.size}, sortField='activo', sortDir=${sortField == 'activo' ? reverseSortDir : 'asc'})}">
                         Estado
                         <span th:if="${sortField == 'activo'}" th:text="${sortDir == 'asc' ? '▲' : '▼'}"></span>
                     </a>
@@ -117,15 +92,15 @@
                 </td>
                 <td data-column="estado" th:text="${usuario.activo ? 'Activo' : 'No Activo'}" th:style="${usuario.activo ? 'color: green;' : 'color: red;'}"></td>
                 <td>
-                    <form th:action="@{/admin/activate_user}" method="post" style="display:inline;" th:if="${!usuario.activo}">
+                    <form th:action="@{/usuario/activate_user}" method="post" style="display:inline;" th:if="${!usuario.activo}">
                         <input type="hidden" th:value="${usuario.id}" name="userId"/>
                         <button type="submit" class="btn btn-success">Activar</button>
                     </form>
-                    <form th:action="@{/admin/delete_user}" method="post" style="display:inline;" th:if="${!usuario.activo}">
+                    <form th:action="@{/usuario/delete_user}" method="post" style="display:inline;" th:if="${!usuario.activo}">
                         <input type="hidden" th:value="${usuario.id}" name="userId"/>
                         <button type="submit" class="btn btn-danger">Rechazar</button>
                     </form>
-                    <form th:action="@{/admin/delete_user}" method="post" th:id="'deleteForm-' + ${usuario.id}" style="display:inline;" th:if="${usuario.activo}">
+                    <form th:action="@{/usuario/delete_user}" method="post" th:id="'deleteForm-' + ${usuario.id}" style="display:inline;" th:if="${usuario.activo}">
                         <input type="hidden" th:value="${usuario.id}" name="userId"/>
                         <button type="button" th:id="'deleteButton-' + ${usuario.id}" class="btn btn-warning" th:onclick="'confirmDelete(' + ${usuario.id} + ')'">Eliminar</button>
                         <button type="button" th:id="'confirmButton-' + ${usuario.id}" class="confirm-button btn btn-danger" th:onclick="'document.getElementById(\'deleteForm-' + ${usuario.id} + '\').submit()'" style="display:none;">Confirmar</button>
@@ -138,25 +113,26 @@
         <nav aria-label="Page navigation">
             <ul class="pagination justify-content-center">
                 <li class="page-item" th:classappend="${currentPage == 0} ? 'disabled'">
-                    <a class="page-link" th:href="@{/admin/show_users(page=${currentPage - 1}, size=${userPage.size}, sortField=${sortField}, sortDir=${sortDir})}" aria-label="Previous">
+                    <a class="page-link" th:href="@{/show_users(page=${currentPage - 1}, size=${userPage.size}, sortField=${sortField}, sortDir=${sortDir})}" aria-label="Previous">
                         <span aria-hidden="true">&laquo;</span>
                     </a>
                 </li>
                 <li class="page-item" th:each="i : ${#numbers.sequence(0, totalPages - 1)}" th:classappend="${i == currentPage} ? 'active'">
-                    <a class="page-link" th:href="@{/admin/show_users(page=${i}, size=${userPage.size}, sortField=${sortField}, sortDir=${sortDir})}" th:text="${i + 1}"></a>
+                    <a class="page-link" th:href="@{/show_users(page=${i}, size=${userPage.size}, sortField=${sortField}, sortDir=${sortDir})}" th:text="${i + 1}"></a>
                 </li>
                 <li class="page-item" th:classappend="${currentPage == totalPages - 1} ? 'disabled'">
-                    <a class="page-link" th:href="@{/admin/show_users(page=${currentPage + 1}, size=${userPage.size}, sortField=${sortField}, sortDir=${sortDir})}" aria-label="Next">
+                    <a class="page-link" th:href="@{/show_users(page=${currentPage + 1}, size=${userPage.size}, sortField=${sortField}, sortDir=${sortDir})}" aria-label="Next">
                         <span aria-hidden="true">&raquo;</span>
                     </a>
                 </li>
             </ul>
         </nav>
     </div>
-    <div class="footer text-center mt-4">
-        <p>&copy; Curso 24/25 Portal Fórmula 1 - Ismael Blázquez</p>
-    </div>
+
+    <div th:replace="~{fragmentos/footer :: footer}"></div>
 </div>
+<script th:src="@{/js/jquery.min.js}"></script>
 <script th:src="@{/js/bootstrap.min.js}"></script>
+<script th:src="@{/js/bootstrap.bundle.min.js}"></script>
 </body>
 </html>
